Break add to schedule modal into component
