--- src/page/Pois/AddWindow.jsx
@@ -1,12 +1,0 @@
-const AddWindow = () => {
-  return (
-    <>
-      <div className="fixed top-0 z-[998] flex h-full w-full flex-row items-center justify-center bg-black opacity-50"></div>
-      <div className="fixed top-0 z-[999] flex h-full w-full flex-row items-center justify-center">
-        <div className="h-40 w-80 rounded-xl bg-white opacity-100"></div>
-      </div>
-    </>
-  );
-};
-
-export default AddWindow;
